Refatorando a exclusão de cozinhas para usar domain services

+++ src/main/java/com/roberto/algafood/infraestructure/repository/CozinhaRepositoryImpl.java
@@ -2,6 +2,7 @@ package com.roberto.algafood.infraestructure.repository;
 
 import com.roberto.algafood.domain.model.Cozinha;
 import com.roberto.algafood.domain.repository.CozinhaRepository;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -34,8 +35,11 @@ public class CozinhaRepositoryImpl implements CozinhaRepository {
 
     @Override
     @Transactional
-    public void remover(Cozinha cozinha) {
-        cozinha = buscar(cozinha.getId());
+    public void remover(Long id) {
+        Cozinha cozinha = buscar(id);
+        if (cozinha == null) {
+            throw new EmptyResultDataAccessException(1);
+        }
         entityManager.remove(cozinha);
     }
